Second attempt WIP
This was my second attempt at making compression that was faster and lossless using pythons built in image library (PIL). This one compresses the image without any corruption and also writes the data to a file with greater speed than before. The issue with this one is that the file is much larger than the image it compresses so there's really no point in using it.